--- src/main/java/pl/project/common/storage/LocalFile.java
@@ -1,22 +1,0 @@
-package pl.project.common.storage;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-public class LocalFile {
-    private String name;
-    private String creationTime;
-    private String lastModified;
-    private Long size;
-    private String downloadUri;
-    private String deleteUri;
-    private String fileType;
-
-}
