+++ src/main/java/pl/project/model/localization/Localization.java
@@ -1,0 +1,66 @@
+
+package pl.project.model.localization;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import lombok.*;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({
+        "ip",
+        "type",
+        "continent_code",
+        "continent_name",
+        "country_code",
+        "country_name",
+        "region_code",
+        "region_name",
+        "city",
+        "zip",
+        "latitude",
+        "longitude",
+        "location"
+})
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Builder
+public class Localization {
+
+    @JsonProperty("ip")
+    private String ip;
+    @JsonProperty("type")
+    private String type;
+    @JsonProperty("continent_code")
+    private String continentCode;
+    @JsonProperty("continent_name")
+    private String continentName;
+    @JsonProperty("country_code")
+    private String countryCode;
+    @JsonProperty("country_name")
+    private String countryName;
+    @JsonProperty("region_code")
+    private String regionCode;
+    @JsonProperty("region_name")
+    private String regionName;
+    @JsonProperty("city")
+    private String city;
+    @JsonProperty("zip")
+    private String zip;
+    @JsonProperty("latitude")
+    private Double latitude;
+    @JsonProperty("longitude")
+    private Double longitude;
+    @JsonProperty("location")
+    private Location location;
+    @JsonIgnore
+    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
+
+}
