+++ src/main/java/pl/project/model/weather/Main.java
@@ -1,0 +1,59 @@
+
+package pl.project.model.weather;
+
+import com.fasterxml.jackson.annotation.*;
+import lombok.*;
+
+import javax.validation.Valid;
+import java.util.HashMap;
+import java.util.Map;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({
+        "temp",
+        "pressure",
+        "humidity",
+        "temp_min",
+        "temp_max",
+        "sea_level",
+        "grnd_level"
+})
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Builder
+public class Main {
+
+    @JsonProperty("temp")
+    private Double temp;
+    @JsonProperty("pressure")
+    private Double pressure;
+    @JsonProperty("humidity")
+    private Integer humidity;
+    @JsonProperty("temp_min")
+    private Double tempMin;
+    @JsonProperty("temp_max")
+    private Double tempMax;
+    @JsonProperty("sea_level")
+    private Double seaLevel;
+    @JsonProperty("grnd_level")
+    private Double grndLevel;
+    @JsonIgnore
+    @Valid
+    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
+
+    @Override
+    public String toString() {
+        return "Main{" +
+                "temp=" + temp +
+                ", pressure=" + pressure +
+                ", humidity=" + humidity +
+                ", tempMin=" + tempMin +
+                ", tempMax=" + tempMax +
+                ", seaLevel=" + seaLevel +
+                ", grndLevel=" + grndLevel +
+                ", additionalProperties=" + additionalProperties +
+                '}';
+    }
+}
