+++ src/main/java/pl/project/model/weather/Rain.java
@@ -1,0 +1,57 @@
+
+package pl.project.model.weather;
+
+import com.fasterxml.jackson.annotation.*;
+
+import javax.validation.Valid;
+import java.util.HashMap;
+import java.util.Map;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({
+        "3h"
+})
+public class Rain {
+
+    @JsonProperty("3h")
+    private Double _3h;
+    @JsonIgnore
+    @Valid
+    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
+
+    public Rain() {
+    }
+
+    public Rain(Double _3h) {
+        super();
+        this._3h = _3h;
+    }
+
+    @JsonProperty("3h")
+    public Double get3h() {
+        return _3h;
+    }
+
+    @JsonProperty("3h")
+    public void set3h(Double _3h) {
+        this._3h = _3h;
+    }
+
+    @JsonAnyGetter
+    public Map<String, Object> getAdditionalProperties() {
+        return this.additionalProperties;
+    }
+
+    @JsonAnySetter
+    public void setAdditionalProperty(String name, Object value) {
+        this.additionalProperties.put(name, value);
+    }
+
+    @Override
+    public String toString() {
+        return "Rain{" +
+                "_3h=" + _3h +
+                ", additionalProperties=" + additionalProperties +
+                '}';
+    }
+}
