--- src/main/java/pl/project/twitter/model/TweetDto.java
@@ -1,23 +1,0 @@
-package pl.project.twitter.model;
-
-import lombok.Builder;
-
-@Builder
-public class TweetDto {
-
-    private String author;
-    private String content;
-    private String accountName;
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public String getAccountName() {
-        return accountName;
-    }
-}
