--- src/main/java/pl/project/weather/controller/WeatherController.java
@@ -1,27 +1,0 @@
-package pl.project.weather.controller;
-
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import pl.project.weather.model.dto.WeatherDto;
-import pl.project.weather.service.WeatherService;
-
-@CrossOrigin()
-@RestController
-@RequestMapping("/api/")
-public class WeatherController {
-    private WeatherService weatherService;
-
-    public WeatherController(WeatherService weatherService) {
-        this.weatherService = weatherService;
-    }
-
-    @GetMapping(value = "weather",
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public WeatherDto getWeather() {
-        return weatherService.getWeather();
-    }
-}
-
