--- src/main/java/pl/project/weather/model/dto/WeatherDto.java
@@ -1,71 +1,0 @@
-package pl.project.weather.model.dto;
-
-import lombok.Builder;
-
-@Builder
-public class WeatherDto {
-
-    private String name;
-    private String weather;
-    private String weatherIcon;
-    private double temp;
-    private double pressure;
-    private double minTemp;
-    private double maxTemp;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getWeather() {
-        return weather;
-    }
-
-    public void setWeather(String weather) {
-        this.weather = weather;
-    }
-
-    public double getTemp() {
-        return temp;
-    }
-
-    public void setTemp(double temp) {
-        this.temp = temp;
-    }
-
-    public double getPressure() {
-        return pressure;
-    }
-
-    public void setPressure(double pressure) {
-        this.pressure = pressure;
-    }
-
-    public double getMinTemp() {
-        return minTemp;
-    }
-
-    public void setMinTemp(double minTemp) {
-        this.minTemp = minTemp;
-    }
-
-    public double getMaxTemp() {
-        return maxTemp;
-    }
-
-    public void setMaxTemp(double maxTemp) {
-        this.maxTemp = maxTemp;
-    }
-
-    public String getWeatherIcon() {
-        return weatherIcon;
-    }
-
-    public void setWeatherIcon(String weatherIcon) {
-        this.weatherIcon = weatherIcon;
-    }
-}
