--- src/main/java/pl/project/weather/model/weather/Clouds.java
@@ -1,57 +1,0 @@
-
-package pl.project.weather.model.weather;
-
-import com.fasterxml.jackson.annotation.*;
-
-import javax.validation.Valid;
-import java.util.HashMap;
-import java.util.Map;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "all"
-})
-public class Clouds {
-
-    @JsonProperty("all")
-    private Integer all;
-    @JsonIgnore
-    @Valid
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    public Clouds() {
-    }
-
-    public Clouds(Integer all) {
-        super();
-        this.all = all;
-    }
-
-    @JsonProperty("all")
-    public Integer getAll() {
-        return all;
-    }
-
-    @JsonProperty("all")
-    public void setAll(Integer all) {
-        this.all = all;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-    @Override
-    public String toString() {
-        return "Clouds{" +
-                "all=" + all +
-                ", additionalProperties=" + additionalProperties +
-                '}';
-    }
-}
