--- src/main/java/pl/project/weather/model/weather/Coord.java
@@ -1,72 +1,0 @@
-
-package pl.project.weather.model.weather;
-
-import com.fasterxml.jackson.annotation.*;
-
-import javax.validation.Valid;
-import java.util.HashMap;
-import java.util.Map;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "lon",
-        "lat"
-})
-public class Coord {
-
-    @JsonProperty("lon")
-    private Double lon;
-    @JsonProperty("lat")
-    private Double lat;
-    @JsonIgnore
-    @Valid
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    public Coord() {
-    }
-
-    public Coord(Double lon, Double lat) {
-        super();
-        this.lon = lon;
-        this.lat = lat;
-    }
-
-    @JsonProperty("lon")
-    public Double getLon() {
-        return lon;
-    }
-
-    @JsonProperty("lon")
-    public void setLon(Double lon) {
-        this.lon = lon;
-    }
-
-    @JsonProperty("lat")
-    public Double getLat() {
-        return lat;
-    }
-
-    @JsonProperty("lat")
-    public void setLat(Double lat) {
-        this.lat = lat;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-    @Override
-    public String toString() {
-        return "Coord{" +
-                "lon=" + lon +
-                ", lat=" + lat +
-                ", additionalProperties=" + additionalProperties +
-                '}';
-    }
-}
