--- src/main/java/pl/project/weather/model/weather/Sys.java
@@ -1,102 +1,0 @@
-
-package pl.project.weather.model.weather;
-
-import com.fasterxml.jackson.annotation.*;
-
-import javax.validation.Valid;
-import java.util.HashMap;
-import java.util.Map;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "message",
-        "country",
-        "sunrise",
-        "sunset"
-})
-public class Sys {
-
-    @JsonProperty("message")
-    private Double message;
-    @JsonProperty("country")
-    private String country;
-    @JsonProperty("sunrise")
-    private Integer sunrise;
-    @JsonProperty("sunset")
-    private Integer sunset;
-    @JsonIgnore
-    @Valid
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    public Sys() {
-    }
-
-    public Sys(Double message, String country, Integer sunrise, Integer sunset) {
-        super();
-        this.message = message;
-        this.country = country;
-        this.sunrise = sunrise;
-        this.sunset = sunset;
-    }
-
-    @JsonProperty("message")
-    public Double getMessage() {
-        return message;
-    }
-
-    @JsonProperty("message")
-    public void setMessage(Double message) {
-        this.message = message;
-    }
-
-    @JsonProperty("country")
-    public String getCountry() {
-        return country;
-    }
-
-    @JsonProperty("country")
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    @JsonProperty("sunrise")
-    public Integer getSunrise() {
-        return sunrise;
-    }
-
-    @JsonProperty("sunrise")
-    public void setSunrise(Integer sunrise) {
-        this.sunrise = sunrise;
-    }
-
-    @JsonProperty("sunset")
-    public Integer getSunset() {
-        return sunset;
-    }
-
-    @JsonProperty("sunset")
-    public void setSunset(Integer sunset) {
-        this.sunset = sunset;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-    @Override
-    public String toString() {
-        return "Sys{" +
-                "message=" + message +
-                ", country='" + country + '\'' +
-                ", sunrise=" + sunrise +
-                ", sunset=" + sunset +
-                ", additionalProperties=" + additionalProperties +
-                '}';
-    }
-}
