--- src/main/java/pl/project/weather/model/weather/Weather.java
@@ -1,102 +1,0 @@
-
-package pl.project.weather.model.weather;
-
-import com.fasterxml.jackson.annotation.*;
-
-import javax.validation.Valid;
-import java.util.HashMap;
-import java.util.Map;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "id",
-        "main",
-        "description",
-        "icon"
-})
-public class Weather {
-
-    @JsonProperty("id")
-    private Integer id;
-    @JsonProperty("main")
-    private String main;
-    @JsonProperty("description")
-    private String description;
-    @JsonProperty("icon")
-    private String icon;
-    @JsonIgnore
-    @Valid
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    public Weather() {
-    }
-
-    public Weather(Integer id, String main, String description, String icon) {
-        super();
-        this.id = id;
-        this.main = main;
-        this.description = description;
-        this.icon = icon;
-    }
-
-    @JsonProperty("id")
-    public Integer getId() {
-        return id;
-    }
-
-    @JsonProperty("id")
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    @JsonProperty("main")
-    public String getMain() {
-        return main;
-    }
-
-    @JsonProperty("main")
-    public void setMain(String main) {
-        this.main = main;
-    }
-
-    @JsonProperty("description")
-    public String getDescription() {
-        return description;
-    }
-
-    @JsonProperty("description")
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    @JsonProperty("icon")
-    public String getIcon() {
-        return icon;
-    }
-
-    @JsonProperty("icon")
-    public void setIcon(String icon) {
-        this.icon = icon;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-    @Override
-    public String toString() {
-        return "Weather{" +
-                "id=" + id +
-                ", main='" + main + '\'' +
-                ", description='" + description + '\'' +
-                ", icon='" + icon + '\'' +
-                ", additionalProperties=" + additionalProperties +
-                '}';
-    }
-}
