--- src/main/java/pl/project/weather/model/weather/WeatherObject.java
@@ -1,94 +1,0 @@
-
-package pl.project.weather.model.weather;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.*;
-
-import javax.validation.Valid;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "coord",
-        "weather",
-        "base",
-        "main",
-        "wind",
-        "rain",
-        "clouds",
-        "dt",
-        "sys",
-        "timezone",
-        "id",
-        "name",
-        "cod"
-})
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Builder
-public class WeatherObject {
-
-    @JsonProperty("coord")
-    @Valid
-    private Coord coord;
-    @JsonProperty("weather")
-    @Valid
-    private List<Weather> weather = null;
-    @JsonProperty("base")
-    private String base;
-    @JsonProperty("main")
-    @Valid
-    private Main main;
-    @JsonProperty("wind")
-    @Valid
-    private Wind wind;
-    @JsonProperty("rain")
-    @Valid
-    private Rain rain;
-    @JsonProperty("clouds")
-    @Valid
-    private Clouds clouds;
-    @JsonProperty("dt")
-    private Integer dt;
-    @JsonProperty("sys")
-    @Valid
-    private Sys sys;
-    @JsonProperty("timezone")
-    private Integer timezone;
-    @JsonProperty("id")
-    private Integer id;
-    @JsonProperty("name")
-    private String name;
-    @JsonProperty("cod")
-    private Integer cod;
-    @JsonIgnore
-    @Valid
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    @Override
-    public String toString() {
-        return "WeatherObject{" +
-                "coord=" + coord +
-                ", weather=" + weather +
-                ", base='" + base + '\'' +
-                ", main=" + main +
-                ", wind=" + wind +
-                ", rain=" + rain +
-                ", clouds=" + clouds +
-                ", dt=" + dt +
-                ", sys=" + sys +
-                ", timezone=" + timezone +
-                ", id=" + id +
-                ", name='" + name + '\'' +
-                ", cod=" + cod +
-                ", additionalProperties=" + additionalProperties +
-                '}';
-    }
-}
