--- src/main/java/pl/project/weather/model/weather/Wind.java
@@ -1,72 +1,0 @@
-
-package pl.project.weather.model.weather;
-
-import com.fasterxml.jackson.annotation.*;
-
-import javax.validation.Valid;
-import java.util.HashMap;
-import java.util.Map;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "speed",
-        "deg"
-})
-public class Wind {
-
-    @JsonProperty("speed")
-    private Double speed;
-    @JsonProperty("deg")
-    private Double deg;
-    @JsonIgnore
-    @Valid
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    public Wind() {
-    }
-
-    public Wind(Double speed, Double deg) {
-        super();
-        this.speed = speed;
-        this.deg = deg;
-    }
-
-    @JsonProperty("speed")
-    public Double getSpeed() {
-        return speed;
-    }
-
-    @JsonProperty("speed")
-    public void setSpeed(Double speed) {
-        this.speed = speed;
-    }
-
-    @JsonProperty("deg")
-    public Double getDeg() {
-        return deg;
-    }
-
-    @JsonProperty("deg")
-    public void setDeg(Double deg) {
-        this.deg = deg;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-    @Override
-    public String toString() {
-        return "Wind{" +
-                "speed=" + speed +
-                ", deg=" + deg +
-                ", additionalProperties=" + additionalProperties +
-                '}';
-    }
-}
