+++ src/main/resources/twitter4j.properties
@@ -1,0 +1,4 @@
+oauth.consumerKey =MUoMq7uUu56u7k23VRn5b7XNN
+oauth.consumerSecret =5geRvV5HX5cyoAo7NppBbFtYdNHbQPmLOUxn8YWSUOsSt7dEZU
+oauth.accessToken =571799416-LcEDzAsOOg6KjFRw7a2VBgnwP0tky5Z5ESaJME34
+oauth.accessTokenSecret =SiM86HIj5nSOCk8SeregdKqy985cFgcIJHhhX0DhicJaU
