+++ src/test/java/pl/project/weather/controller/WeatherControllerTest.java
@@ -7,8 +7,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import pl.project.weather.model.dto.WeatherDto;
-import pl.project.weather.service.WeatherService;
+import pl.project.controller.WeatherController;
+import pl.project.model.dto.WeatherDto;
+import pl.project.service.WeatherService;
 
 import javax.ws.rs.core.MediaType;
