+++ src/test/java/pl/project/weather/model/OpenWeatherApiTest.java
@@ -6,6 +6,7 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
+import pl.project.model.OpenWeatherApi;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
